added lots of css and then an edit / delete channel button

+++ app/controllers/api/channels_controller.rb
@@ -11,10 +11,14 @@ class Api::ChannelsController < ApplicationController
 
     def destroy
         @channel = Channel.find(params[:id]);
-        if @channel.destroy
-            render :show
+        if @channel && @channel.server.admin.id == current_user.id
+            if @channel.destroy
+                render :show
+            else
+                render json: @channel.errors.full_messages, status: 404
+            end
         else
-            render json: @channel.errors.full_messages, status: 404
+            render json: ["You don't have permission for that."], status: 403
         end
     end
 
@@ -22,6 +26,20 @@ class Api::ChannelsController < ApplicationController
         @channel = Channel.find(params[:id]);
         render :show
     end
+
+    def update
+        @channel = Channel.find(params[:id]);
+        if @channel && @channel.server.admin.id == current_user.id
+            if @channel.update(channel_params)
+                render :show
+            else
+                render json: @channel.errors.full_messages, status: 404
+            end
+        else
+            render json: ["You don't have permission for that."], status: 403
+        end
+    end
+
 
     def index
         @channels = Channel.where(server_id: params[:server_id])
